working on v 002

+++ RAC_switch/Logger.cs
@@ -11,18 +11,29 @@ namespace RAC_switch
     {
         public static bool bEnableLogging = true;
         public static string sLogFile="\\RAC_switch.log.txt";
+        public static bool bFirstWrite = true;
+
         public static void WriteLine(string s)
         {
             System.Diagnostics.Debug.WriteLine(s);
             if (bEnableLogging)
             {
                 try{
-                    System.IO.FileInfo fi = new FileInfo(sLogFile);
-                    if (fi.Length > 2000000)
+                    if (System.IO.File.Exists(sLogFile))
                     {
-                        System.IO.File.Move(sLogFile, sLogFile + ".bak");
+                        System.IO.FileInfo fi = new FileInfo(sLogFile);
+                        if (fi.Length > 2000000)
+                        {
+                            System.IO.File.Delete(sLogFile + ".bak");
+                            System.IO.File.Move(sLogFile, sLogFile + ".bak");
+                        }
                     }
                     using(StreamWriter sw = new StreamWriter(sLogFile, true)){
+                        if (bFirstWrite)
+                        {
+                            sw.WriteLine("#### rac_switch "+getVersion()+" ####");
+                            bFirstWrite = false;
+                        }
                         sw.WriteLine(timestamp + "\t" + s);
                     }
                 }catch(Exception ex){
@@ -39,6 +50,27 @@ namespace RAC_switch
                     now.Year, now.Month, now.Day,
                     now.Hour, now.Minute, now.Second);
             }
+        }
+        public static string getVersion()
+        {
+            string s = "001";
+            s = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version.ToString();   // 1.0.0.0
+         
+            return myVersion(s);
+        }
+
+        static string myVersion(string s)
+        {
+            string sIn = s;
+            string sOut="";
+            for (int x = sIn.ToCharArray().Length - 1; x >= 0; x--)
+            {
+                char c = sIn.ToCharArray()[x];
+                if(c!='.')
+                    sOut += c;
+            }
+            return sOut;
+
         }
     }
 }
